+++ src/Scheme.php
@@ -11,9 +11,13 @@ use Uri\Rfc3986\Uri;
  */
 final class Scheme
 {
+    /**
+     * @param ?\WeakReference<Uri|Concrete> $parsed
+     */
     private function __construct(
         private string $value,
-        private bool $less = false,
+        private bool $less,
+        private ?\WeakReference $parsed,
     ) {
     }
 
@@ -28,10 +32,31 @@ final class Scheme
         }
 
         try {
-            return Url::of($value.'://a.org/')->scheme();
+            // this variable is here to keep a reference to the underlying
+            // parsed object
+            $url = Url::of($value.'://a.org/');
+            $self = $url->scheme();
         } catch (\Exception) {
             throw new \DomainException($value);
         }
+
+        /** @psalm-suppress ImpureMethodCall */
+        $parsed = $self->parsed?->get();
+
+        if (\is_null($parsed)) {
+            return $self;
+        }
+
+        /** @psalm-suppress ImpureMethodCall */
+        if (
+            $parsed->getPath() !== '/' ||
+            !\is_null($parsed->getQuery()) ||
+            !\is_null($parsed->getFragment())
+        ) {
+            throw new \DomainException($value);
+        }
+
+        return $self;
     }
 
     /**
@@ -45,12 +70,17 @@ final class Scheme
         /** @psalm-suppress ImpureMethodCall */
         $scheme = $parsed->getRawScheme();
 
+        /** @psalm-suppress ImpureMethodCall */
         return match ($scheme) {
             null => match (\str_starts_with($origin, '//')) {
                 true => self::less(),
                 false => self::none(),
             },
-            default => new self($scheme),
+            default => new self(
+                $scheme,
+                false,
+                \WeakReference::create($parsed),
+            ),
         };
     }
 
@@ -79,7 +109,12 @@ final class Scheme
             return self::none();
         }
 
-        return new self($scheme);
+        /** @psalm-suppress ImpureMethodCall */
+        return new self(
+            $scheme,
+            false,
+            \WeakReference::create($parsed),
+        );
     }
 
     /**
@@ -88,7 +123,7 @@ final class Scheme
     #[\NoDiscard]
     public static function none(): self
     {
-        return new self('');
+        return new self('', false, null);
     }
 
     /**
@@ -99,7 +134,7 @@ final class Scheme
     #[\NoDiscard]
     public static function less(): self
     {
-        return new self('', true);
+        return new self('', true, null);
     }
 
     #[\NoDiscard]