+++ CHANGELOG.md
@@ -13,6 +13,12 @@
 ### Fixed
 
 - Url parsing idempotency with partially encoded characters
+- `Query::of()` silently lost characters
+- `Path::of()` silently lost characters
+- `Host::of()` silently lost characters
+- `Password::of()` silently lost characters
+- `User::of()` silently lost characters
+- `Scheme::of()` silently lost characters
 
 ## 5.3.0 - 2026-05-14